requirements.txt
selenium~=3.141.0
configparser~=5.0.0
pytest~=6.0.1
pytest-html~=2.1.1
allure-pytest~=2.8.17
requests~=2.25.0
urllib3~=1.25.10
Appium-Python-Client~=1.0.2
poium~=1.0.1
deepdiff~=5.5.0
pandas~=1.3.1
jsonpath~=0.82
paho-mqtt